Use `s3fs` instead of `boto3`
