+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 as a connector.
 * Rename `s3_connector.py` to `s3.py` for readability and consistency with the
 rest of the connectors.
+* Now `S3Connector` relies on `s3fs` instead of on `boto3`, and it inherits
+from `FsspecConnector`
 
 ### Features
